update loading and clearing Ember in node tests
