+++ tests/node/helpers/setup-app.js
@@ -1,10 +1,7 @@
 /* eslint-disable no-console */
 
-const path = require('path');
-const distPath = path.join(__dirname, '../../../dist');
-const emberPath = path.join(distPath, 'tests/ember');
-const templateCompilerPath = path.join(distPath, 'ember-template-compiler');
 const SimpleDOM = require('simple-dom');
+const { loadEmber, clearEmber } = require('./load-ember');
 
 /*
  * This helper sets up a QUnit test module with all of the environment and
@@ -60,17 +57,12 @@ const SimpleDOM = require('simple-dom');
 
 module.exports = function(hooks) {
   hooks.beforeEach(function() {
-    let Ember = (this.Ember = require(emberPath));
+    let { Ember, compile } = loadEmber();
+
+    this.Ember = Ember;
+    this.compile = compile;
 
     Ember.testing = true;
-
-    let precompile = require(templateCompilerPath).precompile;
-    this.compile = function(templateString, options) {
-      let templateSpec = precompile(templateString, options);
-      let template = new Function('return ' + templateSpec)();
-
-      return Ember.HTMLBars.template(template);
-    };
 
     this.run = Ember.run;
     this.all = Ember.RSVP.all;
@@ -91,11 +83,7 @@ module.exports = function(hooks) {
   hooks.afterEach(function() {
     this.run(this.app, 'destroy');
 
-    delete global.Ember;
-
-    // clear the previously cached version of this module
-    delete require.cache[emberPath + '.js'];
-    delete require.cache[templateCompilerPath + '.js'];
+    clearEmber();
   });
 };
